requirements.txt
boto3
azure-storage-blob
google-cloud-storage
python-dotenv
pandas
streamlit
